Added FileSystemBackedCache::connect_from_configuration(). Expanded testing thereof.

+++ systems/core/tests/TestsFor_FileSystemBackedCache.php
@@ -3,33 +3,64 @@
   class TestsFor_FileSystemBackedCache
   {
     protected $directory;
-    protected $cache;
     
+    function test_connect( $tester )
+    {
+      $directory = $this->make_test_directory_path($tester);
+      if( $cache = $this->connect($tester, $directory) )
+      {
+        $key       = "a_key";
+        $output    = "lorem ipsum dolor sit amet";
+      
+        $tester->record("set"            , $cache->set($key, $output));
+        $tester->record("get matches set", $cache->get($key) == $output);
+      
+        $cache = FileSystemBackedCache::connect($directory, 0777);
+        if( $tester->record("reconnected", $cache != null) )
+        {
+          $tester->record("get is stable", $cache->get($key) == $output);
+        }
+      }
+      
+      $this->clean_up($directory);
+    }
+
+
     function test_set_and_get( $tester )
     {
-      $cache  = $this->cache;
-      $output = array("a" => array(3, 4, 0), "b" => (object)array("x" => "y"), "c" => 10, "d" => "e");
-      $key    = "first";
-      $path   = $this->make_path($key);
+      $directory = $this->make_test_directory_path($tester);
+      if( $cache = $this->connect($tester, $directory) )
+      {
+         $output = array("a" => array(3, 4, 0), "b" => (object)array("x" => "y"), "c" => 10, "d" => "e");
+         $key    = "b_key";
+         $path   = $this->make_key_path($key, $directory);
+    
+         $tester->record("set"        , $cache->set($key, $output));
+         $tester->record("file exists", file_exists($path));
+         $tester->record("get"        , ($input = $cache->get($key)) != null);
+         $tester->record("get matches", $output == $input);
+         $tester->record("delete"     , $cache->delete($key));
+         $tester->record("file gone"  , !file_exists($path));
+      }
       
-      $tester->record("set"        , $cache->set($key, $output));
-      $tester->record("file exists", file_exists($path));
-      $tester->record("get"        , $input = $cache->get($key));
-      $tester->record("get matches", $output == $input);
-      $tester->record("delete"     , $cache->delete($key));
-      $tester->record("file gone"  , !file_exists($path));
+      $this->clean_up($directory);
     }
     
     
     function test_expiry( $tester )
     {
-      $cache = $this->cache;
-      $key   = "third";
-      $path  = $this->make_path($key);
+      $directory = $this->make_test_directory_path($tester);
+      if( $cache = $this->connect($tester, $directory) )
+      {
+        $key   = "c_key";
+        $path  = $this->make_key_path($key, $directory);
       
-      $tester->record("set with expiry = -1", $cache->set($key, "value", -1));
-      $tester->record("file exists"         , file_exists($path));
-      $tester->record("get returns null"    , $cache->get($key) == null);
+        $tester->record("set with expiry = -1", $cache->set($key, "value", -1));
+        $tester->record("file exists"         , file_exists($path));
+        $tester->record("get returns null"    , $cache->get($key) == null);
+      }
+      
+      $this->clean_up($directory);
     }
 
 
@@ -41,22 +72,51 @@
   
     function configure( $configuration, $tester )
     {
-      $directory = "/tmp/cache_test_" . time();
-      if( is_dir($directory) or @mkdir($directory, 0777, $recursive = true) )
+      $this->directory = "/tmp/cache_test_" . time();
+      return true;
+    }
+    
+    function connect( $tester, $directory, $create_mode = 0777 )
+    {
+      $cache = FileSystemBackedCache::connect($directory, $create_mode);
+      if( $tester->record("connected", $cache != null) )
       {
-        $this->directory = $directory;
-        $this->cache     = new FileSystemBackedCache($directory);
-        
-        return true;
+        if( $tester->record("directory exists", is_dir($directory)) )
+        {
+          return $cache;
+        }
       }
-      else
+      
+      return null;
+    }
+    
+    function clean_up( $directory )
+    {
+      if( is_dir($directory) )
       {
-        $tester->skip("unable to find/create cache directory: " . $directory);
+        if( $dh = opendir($directory) )
+        {
+          while( $name = readdir($dh) )
+          {
+            $path = sprintf("%s/%s", $directory, $name);
+            if( is_file($path) )
+            {
+              @unlink($path);
+            }
+          }
+        }
+
+        @rmdir($directory);
       }
     }
     
-    function make_path( $key )
+    function make_test_directory_path( $tester )
     {
-      return sprintf("%s/%s", $this->directory, $key);
+      return sprintf("%s_%s", $this->directory, $tester->name);
+    }
+    
+    function make_key_path( $key, $directory )
+    {
+      return sprintf("%s/%s", $directory, $key);
     }
   }
